Aligned class names with that in the ML diagram.
Also added Repeatable class.
Renamed some procedure names so that they don't state that they are a member of their class in the name.
For example getDuckPosition within the PlayerDuck class it sufices to call the function getPosition because it is in a class that show what the function gets the position of.